integreate workflow in the main file
