+++ src/main.cpp
@@ -3,6 +3,8 @@
 #include <filesystem>
 #include <fstream>
 #include <map>
+#include <iostream>
+#include <chrono>
 
 #include "config.h"
 #include "database.h"
@@ -106,7 +108,7 @@ Part2Experiment()
     // Write throughput results to a CSV
     std::ofstream file("part2_results.csv");
     file << "data_size_mb,btree_throughput,binary_throughput\n";
-    for (int i = 0; i < btree_throughputs.size(); i++)
+    for (size_t i = 0; i < btree_throughputs.size(); i++)
     {
         int size_mb = 1 << i;  // Powers of 2 (1MB, 2MB, 4MB, ...)
         file << size_mb << "," << btree_throughputs[i] << ","
@@ -145,7 +147,8 @@ Part3Experiment()
         // next_record_mb
         int num_pairs = (next_record_mb - current_data_mb) * 1000000 / 8;
 
-        std::chrono::duration<double> elapsed_put;
+        std::chrono::duration<double> elapsed_put = std::chrono::duration<double>::zero();
+
         // Insert data up to next_record_mb
         for (int j = 0; j < num_pairs; j++)
         {
@@ -236,31 +239,98 @@ Part3Experiment()
         "\nPart 3 experiment complete. Results written to part3_results.csv\n");
 }
 
+// Helper function to measure time for a task
+template <typename Func>
+void MeasureTime(const std::string &taskName, Func &&task) {
+    auto start = std::chrono::high_resolution_clock::now();
+    task();
+    auto end = std::chrono::high_resolution_clock::now();
+    std::chrono::duration<double> elapsed = end - start;
+    std::cout << taskName << " completed in " << elapsed.count() << " seconds.\n";
+}
+
+void DatabaseWorkflow() {
+    std::string db_name = "example_db";
+    size_t memtable_size = 1024 * 1024;  // 1 MB memtable size
+
+    // Step 1: Open the database
+    std::cout << "Opening the database..." << std::endl;
+    Database db(db_name, memtable_size);
+    db.Open();
+
+    // Step 2: Insert 1 million key-value pairs into the database
+    size_t numKeys = 1000000;  // 1 million keys
+    MeasureTime("Inserting 1 million keys", [&]() {
+        for (size_t i = 0; i < numKeys; ++i) {
+            db.Put(i, i * 10);  // Key: i, Value: i * 10
+        }
+    });
+
+    // Step 3: Retrieve random keys using Get
+    std::cout << "Retrieving some keys using Get..." << std::endl;
+    MeasureTime("Retrieving keys", [&]() {
+        for (size_t i = 0; i < 10; ++i) {
+            int key = i * (numKeys / 10);  // Sample evenly spaced keys
+            int value = db.Get(key);
+            std::cout << "Key: " << key << ", Value: " << value << std::endl;
+        }
+    });
+
+    // Step 4: Delete some keys
+    std::cout << "Deleting some keys..." << std::endl;
+    MeasureTime("Deleting keys", [&]() {
+        for (size_t i = 0; i < 10; ++i) {
+            db.Delete(i * (numKeys / 10));  // Sample evenly spaced keys
+        }
+    });
+
+    // Step 5: Scan a range of keys
+    size_t scanStart = numKeys / 2, scanEnd = scanStart + 100;  // Range in the middle of the keys
+    std::cout << "Scanning keys from " << scanStart << " to " << scanEnd << "..." << std::endl;
+    MeasureTime("Scanning keys", [&]() {
+        auto scanResults = db.Scan(scanStart, scanEnd);
+        for (const auto &kv : scanResults) {
+            std::cout << "Key: " << kv.first << ", Value: " << kv.second << std::endl;
+        }
+    });
+
+    // Step 6: Trigger compaction
+    std::cout << "Triggering compaction by adding more keys..." << std::endl;
+    MeasureTime("Inserting additional keys to trigger compaction", [&]() {
+        for (size_t i = numKeys; i < numKeys + 500000; ++i) {  // Add 500,000 more keys
+            db.Put(i, i * 10);
+        }
+    });
+
+    // Verify a key after compaction
+    int testKey = numKeys + 250000;
+    int value = db.Get(testKey);
+    std::cout << "After compaction, Key: " << testKey << ", Value: " << value << std::endl;
+
+    // Step 7: Close and reopen the database to test persistence
+    std::cout << "Closing the database..." << std::endl;
+    db.Close();
+    std::cout << "Reopening the database and verifying data to test persistence" << std::endl;
+    db.Open();
+
+    MeasureTime("Verifying data after reopening", [&]() {
+        for (size_t i = 0; i < 10; ++i) {
+            int key = i * (numKeys / 10);
+            int value = db.Get(key);
+            std::cout << "Key: " << key << ", Value: " << (value != -1 ? std::to_string(value) : "Not found") << std::endl;
+        }
+    });
+
+    // Step 8: Clean up
+    std::cout << "Cleaning up database files..." << std::endl;
+    db.Close();
+    std::filesystem::remove_all(db_name);
+    std::cout << "Database files cleaned up." << std::endl;
+}
+
 int
 main()
 {
-    // Part2Experiment();
-    Part3Experiment();
-
-    // std::filesystem::remove_all("db");
-
-    // Database db("db", MEMTABLE_SIZE);
-
-    // db.Open();
-
-    // // add 300k key-value pairs
-    // printf("Adding 300k key-value pairs to the database...\n");
-    // for (int i = 0; i < 700000; i++)
-    // {
-    //     db.Put(i, i);
-    // }
-
-    // // get 1000 random keys
-    // printf("Getting 1000 random keys\n");
-    // for (int i = 0; i < 1000; i++)
-    // {
-    //     db.Get(rand() % 300000);
-    // }
-
-    // return 0;
+    DatabaseWorkflow();
+    return 0;
 }
